Creating notes, retrieving correct notes when logging in.

+++ src/com/simplenote/android/APIHelper.java
@@ -44,21 +44,23 @@ public class APIHelper {
 			    String key = jsonNote.getString("key");
 			    Date modify = parseDate(jsonNote.getString("modify"));
 			    
-			    if (! (checkNoteExists(key) && ! mDbHelper.checkNewerNote(key, modify))) {		
-			    	if (mDbHelper.checkNewerNote(key, modify)) {
-						if ( Constants.LOGGING ) { Log.i(Constants.TAG, "Note " + key + " is newer on server - retrieving"); }
-			    		mDbHelper.deleteNote(key);
-			    	} else {
-			    		if ( Constants.LOGGING ) { Log.i(Constants.TAG, "Note " + key + " is missing from device - retrieving"); }
-			    	}
-				    
-				    authResponse = APIBase.HTTPGet(Constants.API_NOTE_URL + "?key=" + key + "&auth=" + token + "&email=" + email);
-				    String title = authResponse.resp;
-				    if (title.indexOf('\n') > -1) {
-				    	title = title.substring(0, title.indexOf('\n'));
+			    if (! jsonNote.getString("deleted").equals("true")) {
+				    if (! (checkNoteExists(key) && ! mDbHelper.checkNewerNote(key, modify))) {		
+				    	if (mDbHelper.checkNewerNote(key, modify)) {
+							if ( Constants.LOGGING ) { Log.i(Constants.TAG, "Note " + key + " is newer on server - retrieving"); }
+				    		mDbHelper.deleteNote(key);
+				    	} else {
+				    		if ( Constants.LOGGING ) { Log.i(Constants.TAG, "Note " + key + " is missing from device - retrieving"); }
+				    	}
+					    
+					    authResponse = APIBase.HTTPGet(Constants.API_NOTE_URL + "?key=" + key + "&auth=" + token + "&email=" + email);
+					    String title = authResponse.resp;
+					    if (title.indexOf('\n') > -1) {
+					    	title = title.substring(0, title.indexOf('\n'));
+					    }
+						
+				        mDbHelper.createNote(key, title, authResponse.resp, jsonNote.getString("modify"));
 				    }
-					
-			        mDbHelper.createNote(key, title, authResponse.resp, jsonNote.getString("modify"));
 			    }
 			}	
 		} catch (JSONException e) {
@@ -78,13 +80,13 @@ public class APIHelper {
 		mPrefs.getString("token", null);
 
 		String authBody = APIBase.encode( "email=" + mPrefs.getString("email", "")
-				+ "&password=" + mPrefs.getString("password", ""), true );
+				+ "&password=" + mPrefs.getString("password", ""), true, true );
 		Response authResponse = APIBase.HTTPPost( Constants.API_LOGIN_URL, authBody );
 		
 		if (authResponse.statusCode == 200) { // successful auth login
 			if ( Constants.LOGGING ) { Log.i(Constants.TAG, "Login auth success with API server."); }
 			String token = authResponse.resp;
-			authBody = APIBase.encode(body , true);
+			authBody = APIBase.encode(title + "\n" + body , true, false);
 			authResponse = APIBase.HTTPPost( Constants.API_UPDATE_URL + "?email=" + mPrefs.getString("email", "") 
 					+ "&auth=" + token, authBody);
